refactor eval code and add some new functionalities
